Devoted time and user integration

+++ src/ticket/devoted-time/devoted-time.controler.ts
@@ -2,6 +2,10 @@ import { NextFunction, Request, Response } from 'express'
 
 import { entityManager, entityLoader } from '../../shared/db/orm.js'
 import { DevotedTime } from './devoted-time.entity.js'
+import { decodeToken } from '../../token-decoder.js';
+import { Ticket } from '../ticket.entity.js';
+import { Project } from '../../project/project.entity.js';
+import { error } from 'console';
 
 const em = entityManager;
 
@@ -30,15 +34,23 @@ function sanitizeDevotedTimeInput(req: Request, res: Response, next: NextFunctio
 
 async function findAll(req: Request, res: Response) {
   try {
-    const times = await em.find(
-      DevotedTime,
-      {},
-      { populate: []}
-    )
+    const token = req.headers.authorization
 
-    res
-      .status(200)
-      .json(times)
+    if (token) {
+      const { userId } = decodeToken(token)
+      
+      const times = await em.find(
+        DevotedTime,
+        {user: {id: userId}},
+        { populate: [],
+          orderBy: {date: 'DESC'}
+        },
+      )
+
+      res
+        .status(200)
+        .json(times)
+    }
   } catch(error: any) {
     res
       .status(500)
@@ -48,15 +60,33 @@ async function findAll(req: Request, res: Response) {
 
 async function findTicketsDevotedTime(req: Request, res: Response) {
   try {
-    const ticketId = Number.parseInt(req.params.ticketId)
+    const token = req.headers.authorization
 
-    const times = await em.qb(DevotedTime)
-      .where({ ticket: ticketId})
-      .getResultList();
-    
-    res
-      .status(200)
-      .json(times)
+    if (token) {
+      const { userId, userIsAdmin } = decodeToken(token)     
+      const ticketId = Number.parseInt(req.params.ticketId)
+
+      // Validaciones para no admins
+      if (await noAdminValidation(userIsAdmin, ticketId, userId)) {
+        res.status(403).json({
+            message: 'No posee los permisos necesarios para acceder.'
+          })
+          return
+      }
+
+      const times = await em.find(
+          DevotedTime,
+          {ticket: {id: ticketId}},
+          { populate: [],
+            orderBy: {date: 'DESC'}
+          },
+        )
+      
+      res
+        .status(200)
+        .json(times)
+    }
+
   } catch(error: any) {
     res
       .status(500)
@@ -66,16 +96,30 @@ async function findTicketsDevotedTime(req: Request, res: Response) {
 
 async function findOne(req: Request, res: Response) {
   try {
-    const id = Number.parseInt(req.params.id)
-    const ticketId = Number.parseInt(req.params.ticketId)
+    const token = req.headers.authorization
 
-    const time = await em.qb(DevotedTime)
-      .where({ id: id, ticket: ticketId})
-      .getSingleResult();
+    if (token) {
+      const { userId, userIsAdmin } = decodeToken(token)    
+      const id = Number.parseInt(req.params.id)
+      const ticketId = Number.parseInt(req.params.ticketId)
+      
+      // Validaciones para no admins
+      if (await noAdminValidation(userIsAdmin, ticketId, userId)) {
+        res.status(403).json({
+            message: 'No posee los permisos necesarios para acceder a la entrada.'
+          })
+          return
+      }
+      
+      const time = await em.findOneOrFail(DevotedTime,
+        { id, ticket: ticketId}
+      )
 
-    res
-      .status(200)
-      .json(time)
+      res
+        .status(200)
+        .json(time)
+    }
+    
   } catch(error: any) {
     res
       .status(500)
@@ -85,15 +129,29 @@ async function findOne(req: Request, res: Response) {
 
 async function add(req: Request, res: Response) {
   try {
-    const time = em.create(DevotedTime, req.body);
+    const token = req.headers.authorization
 
-    // delete time.id; //elimino el id que viene desde el front
-
-    await em.flush()
-
-    res
-      .status(201)
-      .json(time)
+    if (token) {
+      const { userId, userIsAdmin } = decodeToken(token)
+      const ticketId = Number.parseInt(req.params.ticketId)
+      
+      // Validaciones para no admins
+      if (await noAdminValidation(userIsAdmin, ticketId, userId)) {
+        res.status(403).json({
+            message: 'No posee los permisos necesarios para crear la entrada.'
+          })
+          return
+      }
+      
+      req.body.newField = 'ticket'
+      req.body.ticket = ticketId
+      const time = em.create(DevotedTime, req.body);
+      await em.flush()
+  
+      res
+        .status(201)
+        .json(time)
+    }
   } catch (error: any) {
     res
       .status(500)
@@ -103,16 +161,30 @@ async function add(req: Request, res: Response) {
 
 async function update(req: Request, res: Response) {
   try {
+    const token = req.headers.authorization
+
+    if (token) {
+      const { userId, userIsAdmin } = decodeToken(token)
+      const ticketId = Number.parseInt(req.params.ticketId)
+      
+      // Validaciones para no admins
+      if (await noAdminValidation(userIsAdmin, ticketId, userId)) {
+        res.status(403).json({
+            message: 'No posee los permisos necesarios para modificar la entrada.'
+          })
+          return
+      }
+      
     const id = Number.parseInt(req.params.id)
     const time = em.getReference(DevotedTime, id)
 
     em.assign(time, req.body)
-
     await em.flush()
 
     res
       .status(200)
       .json({ message: 'Devoted time updated', data: time})
+    }
   } catch(error: any) {
     res
       .status(500)
@@ -122,18 +194,47 @@ async function update(req: Request, res: Response) {
 
 async function remove(req: Request, res: Response) {
   try {
-    const id = Number.parseInt(req.params.id)
-    const time = em.getReference(DevotedTime, id)
+    const token = req.headers.authorization
 
-    await em.removeAndFlush(time)
+    if (token) {
+      const { userId, userIsAdmin } = decodeToken(token)
+      const ticketId = Number.parseInt(req.params.ticketId)
+      
+      // Validaciones para no admins
+      if (await noAdminValidation(userIsAdmin, ticketId, userId)) {
+        res.status(403).json({
+            message: 'No posee los permisos necesarios para eliminar la entrada.'
+          })
+          return
+      }
+    
+      const id = Number.parseInt(req.params.id)
+      const time = em.getReference(DevotedTime, id)
 
-    res
-      .status(200)
-      .json({ message: 'Devoted time removed', data: time})
+      await em.removeAndFlush(time)
+
+      res
+        .status(200)
+        .json({ message: 'Devoted time removed', data: time})
+    }
   } catch(error: any) {
     res
       .status(500)
       .json({ message: error.message})
+  }
+}
+
+async function noAdminValidation(userIsAdmin: boolean, ticketId: number, userId: number) {
+  if (!userIsAdmin) {
+    // Valido que el ticket exista
+    await em.findOneOrFail(Ticket, { id: ticketId})
+
+    // Valido que el usuario esté asignado al proyecto o al ticket
+    const validationProject = await em.findOne(Project, { tickets: {id: ticketId}, users: {id: userId} })
+    const validationTicket = await em.findOne(Ticket, { id: ticketId, responsible: {id: userId} })
+
+    if (validationProject == undefined && validationTicket == undefined) { return true }
+
   }
 }
